Added OtherSQLClients guide doc
In addition to the main examples (e.g. Heidi), I think it's useful to give an overview of other clients we've had experience with. I have a bunch I've tried, so I'll add them here with a similar 'brief summary / positive / negative' format. Mostly for inforamtion, with external links to installation and user manuals.

Definitely not thinking we write full installation guides for all of them